Utilisation de nos icônes si existant dans categories.ts
