+++ app/PlaceSearch/index.tsx
@@ -31,6 +31,7 @@ import SearchHistory from './_components/SearchResults/SearchHistory'
 import SearchLoader from './_components/SearchResults/SearchLoader'
 import SearchNoResults from './_components/SearchResults/SearchNoResults'
 import SearchResultsContainer from './_components/SearchResults/SearchResultsContainer'
+import useIcons from '../effects/useIcons'
 
 /* I'm  not sure of the interest to attache `results` to each state step.
  * It could be cached across the app. No need to re-query photon for identical
@@ -238,6 +239,8 @@ export default function PlaceSearch({
 	const isLoading =
 		!step.results && step.inputValue != null && step.inputValue?.length >= 3
 
+	const icons = useIcons()
+
 	return (
 		<div>
 			<LogoWrapper $sideSheet={sideSheet}>
@@ -347,6 +350,7 @@ export default function PlaceSearch({
 												q: undefined,
 											})
 										},
+										icons,
 									}}
 								/>
 							) : (
